requirements.txt
python-telegram-bot
google-generativeai
google-api-python-client
google-auth-oauthlib
python-dotenv
pytz